\name{nn-class}
\alias{nn-class}
\alias{C++Object-class}
\alias{RcppClass-class}
\title{Neural Network Classes}
\description{
This package uses 'Rcpp' to create modules and functions that encapsulate NNs created using 'nnlib2' (a collection of C++ classes for creating NNs).}
\author{
Vasilis N. Nikolaidis <vnnikolaidis@gmail.com>
}
\note{
Due to their relative simplicity \code{\link{MAM}} and the corresponding 'mam_nn' C++ class are a good starting example of how to implement a new NN using nnlib2 (source code is available on GitHub).}
\seealso{ modules \code{\link{BP}}, \code{\link{MAM}}, \code{\link{LVQs}}, functions \code{\link{Autoencoder}}, \code{\link{LVQu}}.}



